Simplify fixture symbol persistence lifecycle
